プロフィール仮登録処理をhook化
